Add options to add/remove mob spawns from biomes.
This leverages the Biome Modification API to allow users to add or
remove mob spawns from biomes specified in the config. The ENTITY_TYPE
is referenced from the default registry, but as far as I can tell, this
should be okay since the recommended method of registering new entities
is done through this defaulted registry. If an nonexistent entity is
specified, then the world should not load; if a nonexistent biome is
specified, then the modification API silently skips the change.

It's very important that the latest version of Cloth Config is used for
this addition, since it fixes an issue where complex lists did not save.